requirements.txt
tensorflow
numpy
deepface
matplotlib
seaborn
faiss-cpu
scikit-learn
tabulate
